Set up our infrastructure for making calls to the engine
I'm reconsidering how I'm doing this---the double indirection seems
like a code smell.  I know it's intended to keep from having to
unmarshall stuff from the blog alist in each actual xml-rpc call, but
I'm wondering if a macro mightn't be the right actual answer for that.
Or something; it just doesn't sit quite right with me at the moment.