Fix date issues.
Use GMT when sending a post, because it's the only thing that's
unambiguous.

Unfortunately, when WP sends back a date, it doesn't include a
timezone spec, so when xml-rpc goes to interpret it, it presumes it is
GMT and adjusts the result to the local time zone---so the best thing
to do is *read* the _gmt version, but regard it as the local timezone.