Filter the HTML to remove extraneous whitespace.
Doing this while maintaining significant whitespace is tough.  But
WordPress, even when confronted with HTML, has a tendency to do things
like add in extra <br/> tags when it sees newlines and such.  We want
to be able to wrap our content in org-mode and not have that wrapping
imposed on the end result.  I need to put in more tests for corner
cases, but I think this is the way to go about it.